成功部署V2合约到BSC测试网
- 修复所有编译错误
- 部署所有合约成功
- 保存新合约地址到.env
- 合约地址:
  * 多签钱包: 0x534C2c0DF7F06aB6e66E704D4aE809DDa6883737
  * HCF代币: 0x78B7D17C3f98BB47955A155F661f9042F1717288
  * 质押合约: 0x5eA9EDb9446C3FdbAc8ac3BcAf6bF8D8C137D5cD
  * 节点NFT: 0x6fDB1B1F09665Ac00C26701F5E1F92F4652D6F85
  * 交易所: 0x6729c0977325772cF6750eD65f9e3E07f331E104

+++ contracts/HCFMarketControlV2.sol
@@ -234,7 +234,7 @@ contract HCFMarketControlV2 is Ownable, ReentrancyGuard {
         currentProductionMultiplier = PRECISION - reductionRate;
         
         // 计算新的日化收益率
-        uint256[5] memory currentRates = [40, 50, 60, 70, 80];  // 基础0.4%-0.8%
+        uint256[5] memory currentRates = [uint256(40), 50, 60, 70, 80];  // 基础0.4%-0.8%
         uint256[5] memory newRates;
         
         for (uint256 i = 0; i < 5; i++) {
@@ -327,7 +327,7 @@ contract HCFMarketControlV2 is Ownable, ReentrancyGuard {
         uint256 decayPercent = decayUnits * decayRate;  // 每亿减0.1%
         
         // 应用衰减到质押收益率
-        uint256[5] memory currentRates = [40, 50, 60, 70, 80];
+        uint256[5] memory currentRates = [uint256(40), 50, 60, 70, 80];
         uint256[5] memory newRates;
         
         for (uint256 i = 0; i < 5; i++) {
@@ -500,7 +500,7 @@ contract HCFMarketControlV2 is Ownable, ReentrancyGuard {
         currentTaxMultiplier = PRECISION;
         
         // 恢复正常产出
-        uint256[5] memory normalRates = [40, 50, 60, 70, 80];
+        uint256[5] memory normalRates = [uint256(40), 50, 60, 70, 80];
         stakingContract.setBaseDailyRates(normalRates);
         currentProductionMultiplier = PRECISION;
